moved dapregistry module to registry as it now has 2 classes
