+++ capeinfra/pipeline/registry.py
